SDIT-1161 update award now updates changed awards
