+++ src/main/kotlin/uk/gov/justice/digital/hmpps/nomisprisonerapi/adjudications/HearingResultAwardResponse.kt
@@ -5,8 +5,15 @@ import io.swagger.v3.oas.annotations.media.Schema
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @Schema(description = "A list of Hearing result awards created (aka punishment)")
+data class CreateHearingResultAwardResponses(
+  @Schema(description = "an ordered list of award response, the order matching the request order")
+  val awardResponses: List<HearingResultAwardResponse>,
+)
+
+@JsonInclude(JsonInclude.Include.NON_NULL)
+@Schema(description = "A list of Hearing result awards created (aka punishment)")
 data class UpdateHearingResultAwardResponses(
-  @Schema(description = "an ordered list of award response, the order matching the request order for awardRequestsToCreate")
+  @Schema(description = "an ordered list of awards created, the order matching the request order for awardRequestsToCreate")
   val awardResponsesCreated: List<HearingResultAwardResponse>,
   @Schema(description = "a list of awards that were deleted due to this update")
   val awardsDeleted: List<HearingResultAwardResponse>,
